fetch_python/requirements.txt
fastapi
uvicorn
requests
httpx
pytest
python-dotenv
